Added GPLv3 licence
Project is staying here, but only as a reference (so we will develop and release in private) under the most restrictive licence GitHub says it readily offers. Please don't misuse this, otherwise we will simply take the code down. If you're wanting a reference on how we used BLoCs and repositories to get our project up and running in 48 hours with live data, then feel free to browse. Otherwise, please walk on by :)